# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Facturación, Tesorería y Previsión de Caja API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /bank-accounts:
    get:
      operationId: listBankAccounts
      tags: [bank-accounts]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank-accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    patch:
      operationId: updateBankAccount
      tags: [bank-accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /cash-movements:
    get:
      operationId: listCashMovements
      tags: [cash-movements]
      parameters:
        - name: bankAccountId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cash movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashMovement"
    post:
      operationId: createCashMovement
      tags: [cash-movements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashMovement"
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        taxId:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        isUte:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, taxId, isUte, active, createdAt, updatedAt]
    CompanyInput:
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        isUte:
          type: boolean
        active:
          type: boolean
      required: [name, taxId]
    Client:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        taxId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentTermsDays:
          type: ["integer", "null"]
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, active, createdAt, updatedAt]
    ClientInput:
      type: object
      properties:
        companyId:
          type: integer
        name:
          type: string
        taxId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentTermsDays:
          type: integer
        active:
          type: boolean
      required: [companyId, name]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        description:
          type: string
        amount:
          type: string
        taxRate:
          type: ["string", "null"]
        taxAmount:
          type: ["string", "null"]
        total:
          type: string
        expenseDate:
          type: string
        status:
          type: string
        paidAmount:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        supplierName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, description, amount, total, expenseDate, status, createdAt, updatedAt]
    ExpenseInput:
      type: object
      properties:
        companyId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        description:
          type: string
        amount:
          type: string
        taxRate:
          type: string
        expenseDate:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
      required: [companyId, description, amount, expenseDate]
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        bankName:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        currentBalance:
          type: string
        active:
          type: boolean
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, currentBalance, active, createdAt, updatedAt]
    BankAccountInput:
      type: object
      properties:
        companyId:
          type: integer
        name:
          type: string
        bankName:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        currentBalance:
          type: string
        active:
          type: boolean
      required: [companyId, name]
    CashMovement:
      type: object
      properties:
        id:
          type: integer
        bankAccountId:
          type: integer
        type:
          type: string
        amount:
          type: string
        description:
          type: ["string", "null"]
        movementDate:
          type: string
        invoiceId:
          type: ["integer", "null"]
        vendorInvoiceId:
          type: ["integer", "null"]
        bankAccountName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, bankAccountId, type, amount, movementDate, createdAt]
    CashMovementInput:
      type: object
      properties:
        bankAccountId:
          type: integer
        type:
          type: string
        amount:
          type: string
        description:
          type: ["string", "null"]
        movementDate:
          type: string
        invoiceId:
          type: ["integer", "null"]
        vendorInvoiceId:
          type: ["integer", "null"]
      required: [bankAccountId, type, amount, movementDate]
...